change carbon cost from int to float

+++ backend/models/menus.py
@@ -8,7 +8,7 @@ class Menu(BaseModel):
     choices: list[MenuItem]
 
 class CarbonCost(BaseModel):
-    carbon_cost: int
+    carbon_cost: float
     explanation: str
     red_flags: list[str]
     score: int
